requirements.txt
Flask
Flask-SQLAlchemy
PyMySQL
Flask-Login
Werkzeug
python-dotenv
pandas
openpyxl
opencv-python
numpy
